+++ Changelog.md
@@ -3,6 +3,7 @@
 * Hotkeys can be set with a dedicated panel, without going in the settings file.
 * Build orders can be added in a dedicated panel, without having to create a file manually.
 * Builders can be indicated in the resources (optional).
+* Settings and build orders saved in user folder.
 
 ## [1.3.4] - 2022.09.11
 * Resource targets can be split between different resource subtypes, by hovering the mouse.